Combine bulk_discovery and discovery modules
